Updates, including pulling in cookie banner 2.0.6
This also removes an ajax call that was wrappered in a localisation function. Syntax for that type of function has changed in WordPress 5.7+. I initially changed it to that syntax, but after a discussion with Rob, we realised we couldn't test if the AJAX still worked because we're not using it anywhere - so I removed it.